requirements.txt
requests
PyMuPDF
streamlit
pyyaml
requests
flask
matplotlib
openai
google_generativeai
ngrok

